requirements.txt
asn1crypto==1.5.1
cffi==1.15.1
coincurve==17.0.0
cytoolz==0.12.1
eciespy==0.3.13
eth-hash==0.5.1
eth-keys==0.4.0
eth-typing==3.3.0
eth-utils==2.1.0
kcp-py==0.1.3
protobuf==3.20.3
pycparser==2.21
pycryptodome==3.17
toolz==0.12.0
